py_packaging: abstract storage location for resource types
We currently have various types describing Python resources that vary
only in whether data is stored in the filesystem or in memory. I suspect
we can cut down on type overhead by using an enum to represent storage
location.

This commit introduces such an enum on the python resource types.