starlark: implement PythonEmbeddedData type to represent embedded resources
One of the things we sort of lost in the Starlark port was how to write
out files that are used to embed Python in a binary. This was ported in
a very hacky manner by having the generation of the PythonExecutable
do it as a side effect, but only if in a "build artifacts" mode.

This commit aims to start righting the ship by introducing a dedicated
type for representing these embedded resources. This type's build action
will write out these artifact files. So, we can now define a target for
emitting this type and when that target is resolved, these files will be
written. This should open the door for getting rid of our ugly
"build artifacts mode" hack.