project_building: refactor logic for building artifacts
We move a helper function from projectmgmt into project_building. We
refactor it to actually build a resolved target and copy files.
run_from_build() is now implemented in terms of this function.

As part of this, dependency checking should theoretically work again.
Although it may not work in all cases. (It does appear to work for
persisted Rust projects, though.)