project_layout: make pyembed location configurable
We want to support using pyembed from the crates registry, like a normal
Rust package. Let's lay the groundwork for this by allowing the location
of the pyembed crate to be configurable.

Currently, we hardcode the path to the relative path it is always
installed to. But this will change quickly in future commits...