project_layout: pass PyembedLocation into project initialization
The caller is now in charge of where the pyembed crate should be
referenced from.

If running PyOxidizer from its Git checkout, we reference a local
filesystem path to the crate. Otherwise we use the version in the
registry corresponding to the current PyOxidizer version.

There are still some changes to be made here. Release notes will be
added once this transition is done.