project_layout: stop writing copy of pyembed crate
The pyembed crate no longer depends on an unreleased version of
the rust-cpython project's crates. And, we support referencing an
explicit path to the pyembed crate when initializing new projects.

This all means we shouldn't need to vendor the pyembed crate next to
Rust projects any more.

This commit removes support for writing the pyembed files as part of
adding PyOxidizer to a Rust project.