project_building: try to find pyoxidizer.bzl from OUT_DIR
With the pyembed crate gone from local installs, we can no longer rely
on traversing its ancestor directories to find a local pyoxidizer.bzl
file. Instead, we need to rely on a different heuristic.

Cargo build scripts don't receive an environment variable telling them
the path to the project that is building them, unfortunately. But they
do receive the path to the output directory, which is usually in the
directory of the main project being built.