py_packaging: represent Windows embeddable distributions
Building self-contained executables is nice and all. But it creates a
lot of problems (namely the inability to reuse existing Windows binary
wheels). This commit starts to teach PyOxidizer about how to consume
the zip file based Windows embeddable distributions. This will open the
door to utilizing these distributions as part of packaging. There's
still a very long way to go. But the functionality implemented so far
allows us to start consuming these embeddable distributions in other
code paths.