py_packaging: move most of standalone distribution logic to new module
With the introduction of Windows embeddable distributions, we need to
distinguishing between our distribution variants. I'm not yet sure how
this will play out. But we know there will be some common functionality
and some distribution flavor specific functionality. It makes sense to
isolate the generic bits from the specific bits.

This commit moves a bunch of the standalone distribution logic into a
new standalone_distribution module. Some of the moved code will likely
be moved out again. And we likely missed some code that should be moved.
But you have to start somewhere.