py_packaging: refactor importlib derivation to be on PythonDistribution trait
The march towards making a generic to represent Python distributions
continues.

We don't have a clear way to obtain the importlib source for Windows
embeddable distributions. So we defer implementing this for now.

Code has been refactored so derive_importlib() is now generic and can be
used with any distribution, as long as source code and a bytecode
compiler is passed in.