pyembed: add "standalone" crate feature to build without PyOxidizer
I want to make pyembed standalone, without requiring auto-generated
files. This will improve flexibility by allowing anyone to use it as
a library. It will also help with separation of concerns, forcing the
crate providing the executable to bring the embedded resources and
interpreter config, which I think is the better approach, even though
it may be more complicated.