py_packaging: implement ensure_pip()
The Windows embeddable distributions don't contain pip or ensurepip.

So our strategy for installing it is to download a determinstic
version of get-pip.py and run it.

We will need some tweaks to the test utility functions for
obtaining Python distributions because this method will race between
multiple tests once we have multiple tests exercising this
functionality...