py_packaging: securely and deterministically install pip
get-pip.py is not deterministic nor does it validate content hashes
when downloading files from the Internet. See
https://github.com/pypa/get-pip/issues/60. This makes naive usage
inappropriate for PyOxidizer, which wants to ensure downstream
consumers can achieve determinism and isn't the weak link in the
security chain.

Way too much effort was spent developing this commit and figuring
out how to get the packaging tools to install securely and
deterministically. See the long comment in packaging_tool.rs for
details.