py_packaging: install packaging tools in distribution directory
The Windows embeddable distribution is a bit more crippled than
I initially realized. It does not pick up libraries in a
lib or lib/site-packages directory by default. Nor does it seem
to honor PYTHONPATH. The root directory is on sys.path, however.
So we install files directly to the distribution root directory
to enable it to find pip and friends with minimal coercion.