py_packaging: pass exact module names through package filter (#223)
Before, the filter would omit fully qualified module names without a
dot. The new behavior is more inline with user expectations as to how
a filter should work.

As part of this, we add various tests, including a test for
the read_package_root Starlark method, which is how this bug was
reported.

Closed #223.