py_packaging: initial support for linking against shared library
This commit started out as implementing support for linking
Windows embeddable distributions and it still contains some
code for that. But as part of developing it, I realized the
Windows embeddable distributions were too useless to continue
supporting and I pivoted to teaching python-build-standalone
to produce dynamically linked distributions.

This commit teaches the linking and project building code to
handle the presence of an existing libpython shared library.

With this commit, I'm able to build an executable with a Windows
distribution that is dynamically linked. But binaries don't
run because we don't yet install the pythonXY.dll next to the
executable. And we're not handling .pyd extensions properly yet.
This will all be improved in future commits.