py_packaging: install extra files next to built exe
Sometimes exe building needs to install extra files next to the
built binary. This commit introduces a trait member for the exe
builder that allows specifying said files.

We teach the standalone distribution to add pythonXY.dll to the
files so it is copied next to the .exe. The Starlark code for
building an exe is taught to add the returned FileManifest to the
app install manifest.

With this commit, a built exe now runs. However, we're still not
handling .pyd files properly, so some required extension modules
from Python initialization are missing and the process dies pretty
quickly. This will be dealt with in a future commit.