py_packaging: initial support for detecting dynamic extension modules
We want to teach PyOxidizer to use standalone extension module
files from the filesystem, where supported. This commit starts to
implement support for that.

Since not all Python distributions can use existing dynamic
extension modules, we needed to add a mechanism for filtering
collected resources depending on compatibility.

We also had to teach the file resource collection code to handle
extension modules. This is the part that is still a bit hacky.
Future commits will likely need to introduce a new type to represent
this resource.

We also stopped short of fully implementing functionality because
filesystem scanning is resolving the wrong module names on Windows.
So we couldn't accurately test this. Look for improvements in
subsequent commits.