pyembed: move docs for embedded resources into pyembed crate
I believe the docs originally existed in the pyembed crate but were
moved to the Sphinx docs because we weren't publishing the pyembed
crate and couldn't easily link to its generated docs. Now that we
are publishing this crate, this concern doesn't exist and we can
move the docs back to the pyembed crate.

The impetus for doing this is I want to refactor the format of the
embedded resources blob. And it is best for the docs to live in
their final location before doing so.