pyembed: rewrite embedded resources data format
Before, we had 2 different blobs - one for modules and one for resource
data. And we know we want to add new blob types in the future (like
package distribution data and extension modules). Maintaining N
separate data formats would be unwieldy. I've been wanting to
consolidate into a single embedded resources payload for a while.

This commit implements that consolidation.

We invent a new embedded resource data format. We teach pyoxidizer
to write it and pyembed to parse it.

Unlike the previous format, we actually have tests for this new format.
In order to not lose our sanity when writing pyembed tests, the pyembed
crate now depends on pyoxidizer so it can call out to its code for
serializing the data structure.

The new format is much, much better and more extensible. But there's
still a lot we could do. The format won't be set in stone until we
publish a release. Hopefully we can add many of the missing pieces
before we need to worry about backwards compatibility.