embedded_resource: store EmbeddedResourcePythonModule map on EmbeddedPythonResources
We now have a nice type for repsenting a unified Python module. Let's
use it more places.

Logic was refactored as part of the port. But I think I preserved all
important functionality. If not, perhaps we should have more tests!

As part of this we deleted some now-unused resources.