embedded_resource: start to store unified module records
We have a new uber module type to store unified module entries. We
want to start exposing its functionality as part of the embedded
resources.

As the first step towards that, we move storage of module sources
to the new world. Because we want to defer I/O until packaging time,
we had to invent a new type to represent a unified module record
that uses DataLocation instead of Vec<u8>.