embedded_resource: store bytecode in unified modules map
Moving towards a world where we use the uber module type to
represent things instead of separate data structures.

As part of this, we now properly store different bytecode optimization
levels in the embedded resources. Although this may break
pyembed, since it doesn't look in the proper fields when
looking for bytecode.