py_packaging: support adding shared extension modules to embedded resources
The new embedded data structure supports storing this data. And
Windows is capable of loading shared libraries from memory (although
this isn't yet implemented in pyembed). So let's allow dynamically
linked extension modules to be added to the in-memory resources if
the distribution supports it.

We will implement support for loading these extension modules in a
future commit. Also, the .pyd files for the distribution still aren't
embedded. (We don't want to break binary building in the interim.)