pyembed: support storing shared library dependency names in modules entry
In order to better support modules with dependencies, we want to
express shared library dependencies as part of the modules metadata.
This isn't strictly required. But it should enable better debugging
and preemptive loading of missing shared libraries.

This commit teaches our embedded resources data structure to express
the shared library dependencies of an entry.

We do not yet write this field outside of tests and we do not support
doing anything meaningful with it at run time.