pyembed: define blob index similarly to resources index
Unifying the index format is a noble cause on its own. But the
impetus for this is I want to teach blob sections about different
encoding formats and this would necessitate a more complicated
data representation mechanism. Reusing the existing index format
for that seemed the most reasonable thing to do.