pyembed: support NULL padding between elements
I want to design this data format with an eye towards usage outside
of PyOxidizer. One of the requirements of !Rust consumers of this
data structure will be the requirement for null-terminated strings.

This commit introduces the concept of "interior padding" to blob
sections. Interior padding essentially defines how elements are
written out next to each other. By default, there is no padding
add the values "foo" and "bar" are packed like "foobar". But with
null interior padding, we instead write out "foo\0bar\0".