[python-packed-resources] Extract embedded Python resources into own crate
The functionality for reading and writing the embedded resources data
format is generic and doesn't need to live in PyOxidizer or pyembed
itself. While I don't really have the intention of making the
functionality useful outside of PyOxidizer, I think it makes sense to
split it out of PyOxidizer/pyembed because the current split of having
the reader code and writer code in separate crates was kinda wonky.

The new crate is seeded from copies of existing files. Minimal cleanup
was done. The old code still exists and is still being used. We'll move
things to the new crate in a subsequent commit.