python-packed-resources: consolidate resource type via Cow
This commit is way too large and took way too long to hash out.

Essentially, our goal was to unify the type representing resources
so it could work with owned or borrowed fields. This required the use
of Cow and a lot of generics.

One potentially BC change in here is that resources are now
serialized with HashMap instead of BTreeMap. This may introduce
some non-determinism, which is undesirable.