python-packed-resources: add resource flavor field
Previously, we were logically tracking different types of resources
but weren't being explicit about it. This commit introduces an explicit
field in the payload to denote what type of resource is being
represented.

This will open the door to flavor-specific types and validation. Although
I'm not yet sure if that is the direction we'll take.

Having the flavor in the resource itself alleviates the need for the
pyembed crate to maintain its own mapping. So some types have been
removed.