python-packed-resources: add fields to hold relative filesystem paths
This commit teaches our packed resources data structure to encode
relative filesystem paths to resources. Previously, we only supported
defining the resource location via an in-memory payload and
import mechanism.

By allowing resources to be defined as files on the filesystem, we
can maintain a pre-populated mapping of Python resources to filesystem
paths and avoid I/O to discover files on the filesystem. This will cut
down on the number of system calls required to import Python modules.

We do not yet support writing out these new fields in PyOxidizer or
importing modules from paths with the pyembed crate. This support will
occur in subsequent commits.