py_packaging: rename some methods to reflect import location
We're teaching the packed resources data to define modules on the
filesystem. In preparation for exposing this to Starlark, we'll need
to teach APIs for resources about the import location.

The commit renames a bunch of methods to do this. We held off
renaming some methods related to extension modules because the
APIs here are a bit wonky and I want to do them in isolation.