py_packaging: consolidate logic for resolving libpython linking info
The existence of multiple types representing extension modules is
unfortunate. But we can minimize the blast radius by not exposing
the different types beyond where we need to.

This commit establishes a new method for resolving state needed
to link libpython. We're able to move a bit of logic from the
low-level linking code to the EmbeddedPythonResources type.
And we're able to make the fields holding the extension modules
private.