py_packaging: start to support adding path relative Python modules
We've added APIs on the embedded resources type to add source and
bytecode modules. This required a number of refactorings elsewhere,
such as exposing a helper method to resolve filesystem paths and
adding Result to various methods since some called functions can now
fail. I suspect we'll need a bit more refactoring yet as well as many
more tests. But this is a start.