pyembed: consolidate code for retrieving in-memory bytecode
Before, exec_module() didn't respect the current bytecode
optimization level. That was a bug. And get_code() had its own
mechanism for resolving bytecode.

We implement a new helper function for resolving a `memoryview`
for in-memory bytecode and use it for both exec_module() and
get_code(). This makes these implementations vastly easier to
read.