pyembed: refactor resource reading APIs
Like we did for modules, we move a lot of the logic for accessing
low-level resources data into the python_resources module. This makes
the importer code simpler.

As part of this, we refactor how resources data is accessed. Before,
we had an Arc<Box<HashMap<...>>> on each resource reader holding a
reference to the resources for a specific package. In the new world,
we give each resource reader a reference to the importer state and
a package name and we pass the package name through to the new methods
implemented on the resources struct. This will eventually allow the
APIs to return resources from sub-packages, since this seems to be
required for full compatible with the filesystem-based resource
reader.

A side-effect of this change is importer state is now stored as an
Arc<Box<...>> instead of a straight reference. I don't think that
will have any meaningful performance overhead.