python-packed-resources: don't store resources as Arc<Box<...>>
The reason it was done this way was to support the read patterns
formerly required by the pyembed importer. This was a flagrant code
smell and had been bothering me for some time. With the previous
commit to change how resources are accessed, we no longer need the
Arc. So this commit does away with it.