py_packaging: support adding relative path extension modules to embedded resources
Most of the pieces for this were in place already. We did have to
refactor ExtensionModuleData slightly to expose an API to derive
the file path.

We do not yet call the new API. And like in-memory extension module
support, we do not yet support handling library dependencies.
Support will come in time.