starlark: support adding dynamic extension modules
With this commit, we are now able to add extension modules defined
as dynamic libraries to the embedded resources, assuming the current
build configuration supports loading them.

I have tested this with Mercurial on Linux and can confirm that
path-relative .so files are being installed in a location where the
standard file-based importer can find and load them! The
relative path resource entry is being written, making the extension
module available to our custom importer. However, our importer
doesn't yet support importing them from files. This will be
implemented in a subsequent commit.