py_packaging: start to define resources policy
Applications may want to establish policies on where resources can
be loaded from. For example, you may want to require all resources
to be loaded from memory or you may want all resources to be loaded
from the filesystem.

This commit introduces the concept of a "Python Resources Policy."
It allows you to define rules for where resources should be loaded
from.

Currently, we expose the policy to Starlark and thread it through to
our embedded resources type, where the policy is enforced when
resources are added.

In the future, we will teach the binary builders to use the policy
to route resources to the proper resource location.