py_packaging: attach and use resources policy on binary builder
This finishes our initial support for resource policies. We now use
the policy to direct where to add a resource if an explicit location
isn't specified.

There is still some wonkiness around the various add_* APIs. Once we
have parity for location-specific and location-agnostic APIs, this
wonkiness should go away.