py_packaging: refactor broken extensions filtering
I think it is better to prevent the extensions from hitting
the resources state than to filter as part of packaging.

The extensions are only broken for specific distribution flavors.
So we move the logic to the standalone distribution module.

As a benefit of the new approach, we have access to a logger so
we can record the filtering of these extensions. Previously, they
were dropped silently.