embedded_resource: capture link metadata on extension state
There is still room for improving matters here, especially how
various linker options are logically duplicated across fields. But
this was the simplest to implement.

As part of this, EmbeddedPythonResources no longer utilized the
domain-specific extension fields, so they along with code to
populate them has been removed. There is still cleanup to do
elsewhere in the file.