py_packaging: add API for adding relative-path distribution extension module
This is needed to round out the API. As part of this, we also refactor
code that was manually installing dynamically linked extension modules
as part of obtaining the embedded resources instance. Handling these
extension modules as part of producing the binary builder is the proper
place to do this, as distribution extension modules should not come into
play afterwards (unless filtering).