standalone_distribution: use builder APIs for adding distribution resources
Now that we have access to the builder's APIs, we can call its
location agnostic methods for adding resources and let it deal
with the routing logic. This should cut down on code duplication
and help ensure correctness.

We still need some work on the extension APIs. But we can already
see much simpler code.