py_packaging: add a builder interface for adding distribution extension modules
We add a location-agnostic interface for adding distribution extension
modules like exist for most other resource types.

The logic here is quite complicated, as distribution modules are
special (in that they have to be registered in some cases) and the
capabilities of distributions and build configurations varies. I
think the logic as implemented is pretty reasonable. Although it
may require tweaking as bugs / nuances are found.

A side effect of this change is that .pyd-based extensions on
dynamically linked Windows distributions are now loaded from memory
instead of installed relative to the executable. This means fewer
files in the distribution.