embedded_resources: add dynamic depends for filesystem-relative distribution extension modules
On Windows, this will install .dll dependencies next to the .pyd
files, allowing them to be found. I can confirm that
`import sqlite3` now works there.

On other platforms, I suspect we'll need to add a linker path when
building the resulting binary. But the StandaloneDistribution shouldn't
have any shared library dependencies. So we can cross this bridge
later.