embedded_resource: add dynamic library dependency and name to resources
This change should only affect Windows at the moment since it is the
only platform that supports loading extension modules from memory.
After this change, the DLLs for dependency libraries are added to the
embedded resources. The DLL name / resource name is recorded in the
module entry so it can be read at run-time. Although pyembed currently
doesn't read it. So the extension modules still fail to load.