pyembed: use MemoryLoadLibraryEx
We need to use this variant in order to follow library dependencies.
Although we don't yet implement this functionality.

Our default functions should be the same as those implemented by
MemoryModule.c.

As part of implementing this, I discovered some bugs in
memory-module-sys's types. I corrected those in that crate
(I'm the maintainer) and published a new version.

As part of this, we also limit features of the winapi crate so
the build is much, much faster.