pyembed: implement custom LoadLibraryEx callbacks
Previously, our memory DLL loader would only service the
initially loaded module. If that module referenced another
in-memory module, we would fail to locate it and an error
would ensue.

This commit implements custom callbacks for LoadLibraryEx for
module loading, unloading, and symbol lookup.

If a memory module references another memory module, we now
intercept that load request and service it.

When resolving symbols, if the requested module is a memory-loaded
module, we dispatch to the custom symbol resolver.

When unloading a module, we unload a module from memory if the
unloaded module was loaded from there.

Implementing this functionality required storing global state for
memory-loaded modules. We maintain a mapping of module names to
state. The state includes an atomic reference count. The
reference count is needed to avoid freeing modules until the
final reference is released.

The pratical side-effect of this work is that Python extensions
being loaded from memory which depend on other DLLs defined in
memory are now loaded properly. The `_sqlite3` extension in the
Python standard library is one such example. It depends on
`sqlite3.dll`. PyOxidizer was properly packaging both resources
in memory. But before this commit, `sqlite3.dll` could not be
loaded and `import _sqlite3` failed. After this commit, the
dependency chain is followed and `import _sqlite3` works!