py_packaging: rewrite resource file scanning
The resource file scanning code was buggy with regards to handling
of resources in sub-directories and with dots in filenames.
Specifically, it could munge directory separators to dots
inappropriately and result in incorrect resource paths registered
with the importer and spat back out on the filesystem.

This commit sat out to fix this. Along the way, we needed to
drastically change how resources are defined. This gave us an
opportunity to unify the resource structs, like we have been
doing for other Python types.