py_packaging: emit PythonResource from file system scanning
PythonFileResource is mostly redundant with our types in the
resource module.

In order for this to work, we had to implement some missing
conversions and stop normalizing to in-memory variants as
part of the conversion.

I think the normalizing to in-memory had to do with resource
scanning when we are in the context of a temporary filesystem.
So it is possible this introduced some bugs in the form of
race conditions when the temporary directory going away.

If there are race conditions here, callers should be responsible
for converting to in-memory. Although we may need to build out
APIs to better support this...