py_packaging: force resources to be materialized into memory
As hypothesized in a commit message written a few hours ago,
changing the PythonFileResource -> PythonResource conversion to not
force DataLocation::Memory did in fact cause issues with resources
in temporary directories.

This commit adds a new API to PythonResource and its composite
types to force data to be backed by memory. Code in packaging
land calls this to ensure data is backed by memory and not a
temporary directory.