standalone_distribution: support version 5 distributions
We haven't yet published version 5 anywhere stable. So this commit
points to a Linux distribution on my personal web server temporarily.
This will regress macOS and Windows since we now require version 5
distributions.

Most of the code changes here were to update the fields in the JSON.
There were some logic changes that stemmed from removing some
fields. We still have a lot of work to do to utilize the new
fields. But this is a start.