py_packaging: remove supports for Windows embeddable distributions
These proved to be more pain to support than was tolerable.
And things are about to get worse since version 5 of the standalone
distribution metadata exposes a lot of really rich metadata to help
us with packaging.

Some of the code added to support embeddable distributions lingers
because it could be useful someday. But all of the core support code
for this distribution format is removed so we can jettison the
support burden.