py_packaging: pass around bytecode cache tag
This value is now exposed by standalone distributions. So let's
use it in various places.

Most of the code changes stemmed from having to capture the cache
tag on our low-level module types so context was available to
derive the path to the .pyc file for that module.

Buried in these changes was at least one location where
"cpython-37" was hardcoded. This means this change makes us
more agnostic about the specific distribution being used.