distribution: upgrade to latest Python distributions
As part of this, we refactor how distributions are defined.

Before, we had a multiple lists of distributions. Now, we
define a single unified list with a more complex type annotating
the distribution capabilities. This approach - along with
abstracting how the list is accessed - is more flexible and
should allow some more advanced selection of distributions going
forward.