pyembed: use PEP-587 APIs for configuring and initializing interpreter
The new APIs are so, so much better than what came before. Thank you
Victor Stinner!

The new PyConfig struct holds a copy of string data, so we don't
need to have our own type to manage this data and we don't need to
hold a reference for the lifetime of the interpreter.

I tried to port as much functionality to the new API as possible.
And I tried to preserve backwards compatibility as much as possible.
But there is a mismatch between the old and new methods and some
minor regressions were likely introduced. We can mitigate these by
conveging our configuration struct with the new configurations structs
exposed by Python. This will be done in subsequent commits.