pyembed: do not pass marshal and decode_source to setup function
With Python 3.8, it will be possible to initialize our importer
without hacking up the importlib._bootstrap_external frozen
bytecode. But in order to do this, we can't depend on symbols
from that module like we currently do.

This commit removes the dependency on symbols passed from the
hacked up importlib._bootstrap_internal._install() function.

The marshal module is a builtin and can be imported cleanly.
decode_source() is a bit trickier, as it is implemented in Python
in importlib._bootstrap_external. Since we cannot rely on that
module being available, we had to implement it in Python. This
function relies on symbols from the _io module. So we import it
and hold a reference to it for easy calling.