pyembed: establish new types to hold full Python interpreter config
Python 3.8 introduced a ton of new APIs to control interpreter
initialization and configuration.

This commit introduces new types that define the full set of tweakable
configs that now exist.

To reinforce the separation between core Python configs and our
higher-level features, we use different types to represent the different
sets of config options. We keep the existing PythonConfig type
around and teach it to cast into our new type. The new type is
used to configure the Python interpreter.

There's still a bit of work to do here. We will need to teach
Starlark about the new knobs for example. And we'll want to
clean up the higher-level type a bit, adding a Default
implementation, etc. We also likely want to look at the new
run APIs, which are supported in PyConfig. That could probably
simplify our execution code a bit if we could leverage those.