pyembed: run files, code, modules, and repl via Py_RunMain()
This API allows us to run functionality very similarly to how
`python` would do it. Things should be more robust than our
functions. I'm actually tempted to remove our functions,
although I suppose they could be useful. But we may want to
factor them out of the interpreter type, as they shouldn't need
its state to work.

As part of this, there was some refactoring to internals. We now
use a PathBuf to hold the file to run. And we track the interpreter
state via an enum.