pyembed: use non-'static lifetimes for resources data
We just taught MainPythonInterpreter to store a reference to
resources and the importer to store a pointer so that we can use
non-'static lifetimes for resources data.

This commit finishes the support by teaching the configuration
types to store a non-'static lifetime for the resources data.

With this change in place, it should be possible to back resources
data by more dynamic mechanisms, such as loading resources from a
file. However, we do not yet test this.