pyembed: expose indexed resources to Python
We want Python code to be able to query and manipulate resources
indexed by our finder instance. The first step of this is to expose
a read-only interface to allow inspection.

This commit introduces a new Python type to represent indexed
resources. We also expose a method on the finder to return a list
of them.