pyembed: remove bytecode field; clean up file reading
There was a noble purpose in the bytecode field serving as a
cache for bytecode. However, I don't think it was hit that often
because ImportablePythonModule instances aren't typically reused.
For example, when we call find_spec() and exec_module(), we operate
on separate ImportablePythonModule instances.

Let's remove complexity until we have a performance use case
to justify it.

While I was here, I noticed that we were returning a PyMemoryView
backed by stack allocated memory. This is almost certainly a bug
and could lead to crashes when reading bytecode data from the
filesystem.

I changed it to return a PyBytes that owns its memory.