oxidized-importer: new Rust project for producing an extension module
Recent work has made the OxidizedFinder usable outside the context of
PyOxidizer. Let's create a Rust project that allows us to build just
the importer from the pyembed crate.

This is still very much a work in progress. But I can confirm I'm
able to import the produced .so from a standalone `python` process
and access symbols. Progress.