python-packaging: store relative path module source as DataLocation
I think this makes more sense since it is consistent with other
fields. In order to implement this, we had to dynamically derive
the FileManifest for extra files. Hopefully once all fields are
converted, we no longer need to have an explicit extra_files field
and can derive everything from the PrePackagedResource instances.