python-packaging: allow bytecode to be stored verbatim
Before, we always stored the source code of bytecode and would compile
source to bytecode as part of preparing the instance. In many cases
we already have usable bytecode sitting around.

This commit changes how bytecode is stored to allow a variant where
we already have the bytecode. We do not yet have an API for storing
the new variant, however. But the data access path should work just
fine.