oxidized-importer: support building on macOS
This required a cargo config file to make the linker happy. We also
had to teach setup.py about .dylib files. For convenience, we add
a script to build a wheel from macOS. Although I've only tested
this on my local machine, not in CI yet.

And minor changes to setup.py to find the Rust shared libraryu