python_distributions: upgrade to latest python-build-standalone distributions
Because we now produce a shared libpython on Linux and macOS, this
invalidates some of the logic around extensions management since it
was keying off the presence/absence of a shared library. The old logic
was a bit suspect and IMO was worthy of regressing. The plan is to
temporarily eat the test regressions then fix forward.