pyembed: implement OxidizedFinder.iter_modules()
This optional method is used by `pkgutil` to iterate over known
modules.

The test coverage isn't as comprehensive as it could be. But I'm
reasonably confident this works as it should for most scenarios.
We've documented the caveats with our implementation, which I hope
won't be significant for most consumers.

Closes #253.